ml_algo_back_end/requirements.txt
Django==3.1.5
django-cors-headers==3.6.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
scikit-learn==0.24.0
scipy==1.6.0
numpy==1.19.5
pandas==1.2.0
